requirements.txt
# requirements.txt
# Основные зависимости для проекта сегментации транспортных средств

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
albumentations>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0

# Тестирование
pytest>=7.0.0

#Инструменты стиля
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Опционально (для логирования и загрузки данных)
# tensorboard>=2.12.0
# kaggle>=1.5.0